requirements.txt
python-dotenv==0.14.0
requests==2.24.0
terminaltables==3.1.0
